requirements.txt
numpy
pandas
scikit-learn
catboost
joblib
ta
python-binance
